Реализовал RepositoryImpl, проверил напрямую из activity.
