+++ app/src/main/java/com/example/githubusers/utils/Const.kt
@@ -1,0 +1,5 @@
+package com.example.githubusers.utils
+
+//API
+const val BASE_URL_API_GIT_HUB_USERS = "https://api.github.com/"
+const val END_POINT_API_GIT_HUB_USERS = "users"
