--- app/src/main/java/com/example/githubusers/utils/const.kt
@@ -1,4 +1,0 @@
-package com.example.githubusers.utils
-
-const val END_POINT_API_GIT_HUB_USERS = "users"
-const val BASE_URL_API_GIT_HUB_USERS = "https://api.github.com/"
